Extract HTML secondary panes renderer
