+++ docs/DECISIONS.md
@@ -158,6 +158,7 @@
 - Static HTML tool-card rendering lives in `WorkspaceHTMLToolCardRenderer.swift`. `WorkspaceHTMLRenderer.swift` remains the composition point for the Playwright/static harness, while artifact chips, text previews, image/document previews, raw detail blocks, and copy labels stay beside the tool-card HTML renderer. Shared HTML escaping and execution-context chip markup live in `WorkspaceHTMLPrimitives.swift` so terminal rows and tool cards cannot drift.
 - Static HTML terminal rendering lives in `WorkspaceHTMLTerminalRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates terminal pane/entry markup and status-class mapping, while `WorkspaceHTMLPrimitives` remains the single source for escaping and execution-context chip markup shared by terminal rows and tool cards.
 - Static HTML browser rendering lives in `WorkspaceHTMLBrowserRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates browser pane, preview, snapshot, outline, text snippet, and comment markup while staying responsible for only the whole-workspace harness composition.
+- Static HTML secondary pane rendering lives in `WorkspaceHTMLSecondaryPaneRenderer.swift`. `WorkspaceHTMLRenderer.swift` delegates Extensions, Memories, Activity, and Automations pane markup while MCP extension chips, memory cards, activity sections, automation actions, and secondary count labels stay beside the secondary-pane HTML renderer.
 - Browser comment, snapshot, and browser panel state live in `QuillCodeBrowserSurface.swift`; browser address resolution lives in `WorkspaceBrowserLocationResolver.swift`; and browser history/page/comment state transitions live in `WorkspaceBrowserEngine`. The workspace model owns async fetch orchestration, `lastError`, and top-bar refreshes, while reusable browser state contracts, URL-resolution rules, and pure state transitions stay focused and guarded by parity tests.
 - Review comment payload state and event planning live in `WorkspaceReviewCommentPlanner.swift`, not in `WorkspaceModel.swift`. The planner owns path/text trimming, visible review-file validation, line-range normalization, line-kind checks, summary copy, and payload encoding. The workspace model owns only selected-thread validation, event append, persistence, and top-bar refresh.
 - Agent tool execution override composition lives in `WorkspaceToolExecutionOverrideCombiner.swift`, not in `WorkspaceModel.swift`. The combiner owns the precedence order Plan, Remote Project, Browser, Computer Use, Memory, MCP, while the workspace model only creates each optional executor and assigns the combined override to the runner.
